# backend/docs/MilStock_OpenAPI.yaml
openapi: 3.0.0
info:
  title: MilStock API
  version: 1.0.0
  description: Core OpenAPI specification for the MilStock backend.
servers:
  - url: http://localhost:5001/api
    description: Local
  - url: "[ADD_RENDER_BACKEND_URL_HERE]/api"
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [admin, unit, supplier]
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListUsersResponse"
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated
    patch:
      tags: [Users]
      summary: Partially update user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
  /users/{id}/status:
    patch:
      tags: [Users]
      summary: Activate or deactivate user
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, inactive]
      responses:
        "200":
          description: Updated
  /users/{id}/password:
    patch:
      tags: [Users]
      summary: Reset user password
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  example: Password123!
      responses:
        "200":
          description: Password reset
  /products:
    get:
      tags: [Products]
      summary: List products
      responses:
        "200":
          description: Products
    post:
      tags: [Products]
      summary: Create product
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Created
...
  /warehouses:
    get:
      tags: [Warehouses]
      summary: List warehouses
      responses:
        "200":
          description: Warehouses
    post:
      tags: [Warehouses]
      summary: Create warehouse
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Warehouse"
      responses:
        "201":
          description: Created
...
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - in: query
          name: request_type
          schema:
            type: string
            enum: [warehouse_request, supplier_request, warehouse_transfer, provider]
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Orders
    post:
      tags: [Orders]
      summary: Create request/order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created
...
  /movements:
    get:
      tags: [Movements]
      summary: List movement logs
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, completed, cancelled, failed]
        - in: query
          name: movement_type
          schema:
            type: string
      responses:
        "200":
          description: Movements
    post:
      tags: [Movements]
      summary: Create movement
      responses:
        "201":
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: string
      example: 64f000000000000000000001
...
  schemas:
...
    RegisterRequest:
      type: object
      required: [name, email, password, phone, military_number]
      properties:
        name:
          type: string
          example: Unit User
        email:
          type: string
          example: unit@milstock.local
        password:
          type: string
          example: Password123!
        phone:
          type: string
          example: "+966550000000"
        military_number:
          type: string
          example: EMP-001
        role:
          type: string
          enum: [admin, unit, supplier]
...
    User:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        military_number:
          type: string
        role:
          type: string
          enum: [admin, unit, supplier]
        status:
          type: string
          enum: [active, inactive]
        assigned_warehouse:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/Warehouse"
            - nullable: true
    CreateUserRequest:
      allOf:
        - $ref: "#/components/schemas/RegisterRequest"
        - type: object
          properties:
            status:
              type: string
              enum: [active, inactive]
            assigned_warehouse:
              type: string
              nullable: true
    ListUsersResponse:
      type: object
      properties:
        success:
          type: boolean
        count:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
    Product:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
          example: Bread
        quantity:
          type: number
        unit:
          type: string
          enum: [kg, g, liter, Tons, piece, box]
        category:
          type: string
        min_quantity:
          type: number
        warehouse_id:
          type: string
        alert_settings:
          $ref: "#/components/schemas/ProductAlertSettings"
    ProductAlertSettings:
      type: object
      properties:
        low_stock_threshold:
          type: number
          example: 50
        critical_stock_threshold:
          type: number
          example: 10
        expiration_warning_days:
          type: number
          example: 30
        critical_expiration_days:
          type: number
          example: 7
...
    Warehouse:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        code:
          type: string
        location:
          type: string
        capacity:
          type: number
        status:
          type: string
          enum: [active, inactive]
...
    CreateOrderRequest:
      type: object
      required: [request_type, items]
      properties:
        request_type:
          type: string
          enum: [warehouse_request, supplier_request]
        source_warehouse:
          type: string
        destination_warehouse:
          type: string
        supplier_id:
          type: string
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity]
            properties:
              product_id:
                type: string
              quantity:
                type: number
        notes:
          type: string
...